switched int to long for qty, price, value

+++ pe_common.h
@@ -37,7 +37,7 @@ typedef struct{
     dyn_array* product_list;
     dyn_array* traders;
     int num_products;
-    int fee;
+    long fee;
     dyn_array* sigusr_pids;
 } exchange;
